[ARTIFY-27] Introduce Dashboard table, and index and foreign key
