+++ models/dashboard_test.go
@@ -1,0 +1,7 @@
+package models
+
+import "testing"
+
+func Test_Dashboard(t *testing.T) {
+	t.Fatal("This test needs to be implemented!")
+}
